move AuthNavbar to folder molecules
