+++ src/components/Navbar.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import AuthNavbar from './AuthNavbar';
+import AuthNavbar from './molecules/AuthNavbar';
 
 function Navbar() {
   const [navbarOpen, setNavbarOpen] = React.useState(false);
@@ -40,7 +40,6 @@ function Navbar() {
               <Link to="/about" className="flex items-center px-3 py-2 hover-link duration-300">About Us</Link>
             </li>
             <AuthNavbar />
-
           </ul>
         </div>
       </nav>
